*: drive center from route
This was a huge lift, unfortunately, and we seem to have lost a lot of
what is beneficial about layouts, suspense, etc. In the previous
iteration we had a header and footer that were part of the default
layout and would only get rendered once. This is great for really static
sites, but since we need to navigate to a fully "new" site every time
the center changes, we can't use any of those utilities. I imagine if
ISR works as expected the user won't notice any worse behavior.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>