middleware: remove Payload query
We need the middleware to be as fast as possible, as it is in the
critical path for every request. Instead of bothering to dynamically
look up tenants from the CMS, we can hard-code the values we need. The
list of avalanche centers in the US or their web properites is unlikely
to change, so we do not risk correctness issues.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>