package.json: run prettier after code generation
The code generators do not output code perfectly to the Prettier
specification, so we can run Prettier after generation to minimize the
diff. Having to remember to do this every time is annoying, and we know
we need to do it every time, so we might as well just inline the task.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>