*: DRY out server URL logic
We inherited a bit of a mess from the website template. This commit DRYs
it out (while leaving the client-side URL for now). We bridge the gap
between our pipeline-set `$SERVER_DOMAIN` and all of the `$VERCEL*URL`
variables, which only hold the domain, and the
`$NEXT_PUBLIC_SERVER_URL`, which holds the full URL.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>