seed: pass references by ID, not object
We have a couple of collections like `User` with fields that populate
post-creation, like `saveToJwt: true` on fields, or references, etc.
What we want to be able to do with the hash is capture the inputs that
we're using to create the object. When we were putting in the full
tenant, we might put in some minimal fields the first time seeding, but
on re-seed when we look up the tenant we will get more fields, computed
fields, etc. Without adding logic to prune out every type of computed
field, using `tenant.id` instead of `tenant` was the simplest way I
could figure to keep the set of input parameters constant so the hash
would not change. This has the downside of being fragile as deleting all
the tenants and starting over in some other order would change which ID
maps to which tenant, but this script will be very useful to us even with
that caveat. If you're doing big changes like that you can always blow
away the DB and seed from scratch.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>