seed: add an incremental seed endpoint, upsert functionality
We can store a hidden field on our collections that contains the hashed
inputs used to create the object, so a re-seed only needs to update the
(few) items that really changed. The current iteration does not remove
collection items not specified in the newest seed, but that can be
easily added if we find it to be necessary. On my machine, this brings
incremental seed times down to <5s.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>