Pages: look up tenant during revalidation if necessary
It looks like revalidation, when done during publication via the admin
web UI, does not pass the whole tenant in to the hook, even though
creation of the document during seeding does. We can look up the tenant
if it's not set, as it doesn't seem like there's a configuration option
to dictate the depth for the objects passed to hooks.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>